add the IsKnown() method
Use this method to detect if a cluster with
this name already exists when deleting creating or
fetching kubeconfig path.

Add '--force' flag to 'delete' that can be used to
ignore the validation check if a cluster already exists.

Move the message for 'create' from 'pkg' to 'cmd'.
Add message for 'delete' in 'cmd'.