docker: Remove unneeded quotes from the inspect format string
I'm guessing those quotes come from the command line examples where you have to
quote the go template:

  $ docker inspect --format='{{.LogPath}}' $INSTANCE_ID

However, we directly give the array of strings for the exec call so we don't
need to use the shell quotes here.

This removes the need to trim the quotes at the caller sites.