detect docker host IP based on the default route
we were using the well known name host.docker.internal to get the
host IP address, however, it seems that in some systems it resolves
to a loopback address. We use this address to replace the embedded
docker dns IP address in the resolv.conf.
We can't use a loopback address, because coredns obtains the upstream
dns resolvers from the resolv.conf, and evidentelly, it crashes if the
upstream DNS resolver is a loopback, because it meains that is itself,
and that is a loop.