entrypoint: detect cgroupns for cgroup v1
In case the KIND container is run on a cgroup v1 host with --cgroupns=private,
there is no need to do any dances with cgroups.

Detect this and skip cgroup setup.

Otherwise, print a warning that the host cgroupns support is not
enabled, encouraging users to do so. Add a comment to the text to
help future generation of kind maintainers to figure out why this
code is needed.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>