e2e-k8s.sh: support alpha features
FEATURE_GATES and RUNTIME_CONFIG can be set to populate the corresponding kind
config and thus modify the cluster.

For the sake of simplicity, not attempt is made to merge the settings from
GA_ONLY with these new env variables. If GA_ONLY=true is set, the new env
variables are ignored.