Support multiple random ExtraPortMappings
When the default value, 0, is used for the HostPort in ExtraPortMappings
then Kind will determine a random HostPort to use for this mapping. The
validation only allowed a single instance of such a mapping, but now allows
multiple.

Also delay the closing of the randomly determined port until all random
ports have been determined to ensure the same port cannot be returned
multiple times by the operating system.