docs: Fix quick start formatting issues
There were a couple problems in the Quick Start page under the Settings
for Docker Desktop section. There was a broken link to one of the macOS
Docker Desktop settings screenshots. There were also a couple missing
newlines that caused the formatting to be slightly off, running the text
in to the image rather than having a clean separation between them.

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>