Fix site link to LoadBalancer page
The current rendered site does not correctly link local in-repo
documents. This results in the LoadBalancer links on the Ingress User
Guide pages going to a 404.

Unfortunately this behavior is different between "normal" markdown
rendering in local IDEs and while viewing source files in GitHub. But
this change makes the rendered public site correctly link to the page.

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>